+++ src/jvmMain/java/instructions/register/HLTRegister.java
@@ -8,7 +8,7 @@ import main.Registers;
 public class HLTRegister extends Instruction {
 
     public HLTRegister() {
-        super("HLT.R", (short) 0xEE, 2);
+        super("HLT.R", (byte) 0xEE, 2);
     }
     public void execute(Registers registers, Memory memory, Short op) {
